+++ packages/app-page-builder/src/admin/components/Table/Header/TableActions/styled.tsx
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const Container = styled("div")`
+    margin-right: 8px;
+`;
