+++ packages/app-page-builder/src/admin/components/Table/Header/Title/Title.tsx
@@ -1,0 +1,17 @@
+import React, { ReactElement } from "react";
+
+import { Skeleton } from "@webiny/ui/Skeleton";
+
+import { Name } from "./styled";
+
+export interface TitleProps {
+    title?: string;
+}
+
+export const Title = ({ title }: TitleProps): ReactElement => {
+    return (
+        <Name use={"headline6"} tag={"h1"}>
+            {title || <Skeleton />}
+        </Name>
+    );
+};
